+++ components/Pages/ManageLiquidity/WithdrawForm.tsx
@@ -5,10 +5,10 @@ import AssetInput from 'components/AssetInput';
 import { FC, useEffect, useState, useMemo } from 'react';
 import { Controller, useForm } from "react-hook-form";
 import { Asset } from 'types/blockchain';
-import { TokenItemState } from '../ManageLP/lpAtoms';
+import { TokenItemState } from './lpAtoms';
 import { useMultipleTokenBalance } from 'hooks/useTokenBalance';
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { tokenLpAtom } from '../ManageLP/lpAtoms';
+import { tokenLpAtom } from './lpAtoms';
 import { walletState } from 'state/atoms/walletAtoms';
 import { TxStep } from 'hooks/useTransaction';
 import { fromChainAmount } from "libs/num";
@@ -59,7 +59,7 @@ const WithdrawForm = ({ poolId, tokenA, connected }: Props) => {
         else if (tx?.buttonLabel)
             return tx?.buttonLabel
         else
-            return 'Withdraw Liquidity'
+            return 'Withdraw'
 
     }, [tx?.buttonLabel, connected, token?.amount])
