+++ components/Pages/NewPosition/hooks/useTransaction.tsx
@@ -169,8 +169,8 @@ export const useTransaction = ({
         setTxHash(data.transactionHash)
         onBroadcasting?.(data.transactionHash)
         toast({
-          title: 'Swap Success.', 
-          description:  <Finder from={tokenA.symbol}  to={tokenB.symbol} txHash={data.transactionHash} chainId={client.chainId} /> ,
+          title: 'Add Liquidity Success.', 
+          description:  <Finder txHash={data.transactionHash} chainId={client.chainId} > TxHash:  </Finder> ,
           status: 'success',
           duration: 9000,
           position: "top-right",
