+++ components/Pages/Swap/hooks/createSwapMsg.ts
@@ -1,13 +1,13 @@
 // import { toBase64 } from "@arthuryeti/terra";
-// import { LCDClient, Coin, MsgExecuteContract } from "@terra-money/terra.js";
+import { LCDClient} from "@terra-money/terra.js";
 // import { Coin } from 'hooks/useCosmWasmClient'
 import { Coin } from '@cosmjs/launchpad'
 import { coin, EncodeObject } from "@cosmjs/proto-signing";
 
-import { isNativeAsset, toAsset, createAsset } from "./asset";
+import { isNativeAsset, toAsset, createAsset } from "services/asset";
 
-import { Route, SimulationResponse, ReverseSimulationResponse } from "./types";
-import { createExecuteMessage } from '../util/messages';
+import { Route, SimulationResponse, ReverseSimulationResponse } from "types";
+import { createExecuteMessage } from 'util/messages';
 
 export const toBase64 = (obj: object) => {
   return Buffer.from(JSON.stringify(obj)).toString('base64')
