+++ components/Pages/Swap/hooks/useSwap.ts
@@ -1,16 +1,17 @@
 import { useMemo } from "react";
 import { useRecoilValue } from 'recoil';
 import { walletState } from 'state/atoms/walletAtoms';
-import { minAmountReceive } from "./helpers";
+import { minAmountReceive } from "hooks/helpers";
 import { useTokenInfo } from 'hooks/useTokenInfo';
 import { toChainAmount } from "libs/num";
 import { useQueryMatchingPoolForSwap } from 'queries/useQueryMatchingPoolForSwap';
-import { createMsg, createSwapMsgs } from './monoSwap';
-import useSimulate from "./useSimulate";
+import { createMsg, createSwapMsgs } from './createSwapMsg';
+import useSimulate, {Simulated} from "./useSimulate";
 import useTransaction from "hooks/useTransaction";
 import { tokenSwapAtom } from "components/Pages/Swap/swapAtoms";
 import { slippageAtom } from 'components/Pages/Swap/swapAtoms'
 import { fromChainAmount } from "libs/num";
+
 
 const useSwap = ({ reverse }) => {
     const [swapTokenA, swapTokenB] = useRecoilValue(tokenSwapAtom)
@@ -19,15 +20,12 @@ const useSwap = ({ reverse }) => {
     const tokenB = useTokenInfo(swapTokenB?.tokenSymbol)
     const [matchingPools] = useQueryMatchingPoolForSwap({ tokenA, tokenB })
     const slippage = useRecoilValue(slippageAtom)
-
-    // const slippage = "0.05"
-    // const reverse = false
     const token = tokenA?.token_address
     const denom = tokenA?.denom
     const amount = swapTokenA?.amount > 0 ? toChainAmount(swapTokenA?.amount) : ''
     const swapAddress = matchingPools?.streamlinePoolAB?.swap_address || matchingPools?.streamlinePoolBA?.swap_address
 
-    const simulated = useSimulate({
+    const simulated : Simulated = useSimulate({
         client,
         token: reverse ? tokenB?.token_address : tokenA?.token_address,
         amount: reverse ? swapTokenB?.amount > 0 ? toChainAmount(swapTokenB?.amount) : '' : swapTokenA?.amount > 0 ? toChainAmount(swapTokenA?.amount) : '',
@@ -35,8 +33,6 @@ const useSwap = ({ reverse }) => {
         enabled: !(swapTokenA?.tokenSymbol === swapTokenB?.tokenSymbol),
         reverse
     })
-
-
 
 
     const minReceive = useMemo(() => {
