--- hooks/types/index.ts
@@ -1,3 +1,0 @@
-export * from "./cw20";
-export * from "./local";
-export * from "./terraswap";
