+++ pages/404.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import Head from "next/head";
+import Page from "components/Pages/Error";
+import { NextPage } from "next";
+
+const Custom404Page: NextPage = () => {
+  return (
+    <>
+      <Head>
+        <title>Error Page</title>
+      </Head>
+      <Page statusCode={404} />
+    </>
+  );
+};
+
+export default Custom404Page;
