+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import { ErrorBoundary } from 'components/ErrorBoundary'
 import type { AppProps } from 'next/app'
-import { useEffect, FC} from 'react'
+import { useEffect, FC } from 'react'
 import { Toaster } from 'react-hot-toast'
 import { QueryClientProvider } from 'react-query'
 import { RecoilRoot } from 'recoil'
@@ -10,31 +10,37 @@ import { useState } from 'react'
 import theme from "../theme";
 import AppLayout from 'components/Layout/AppLayout'
 import AppLoading from 'components/AppLoading'
+import Head from "next/head";
 
 
-const MyApp:FC<AppProps> = ({ Component, pageProps }) => {
+const MyApp: FC<AppProps> = ({ Component, pageProps }) => {
   const [mounted, setMounted] = useState<boolean>(false);
 
   useEffect(() => setMounted(true), []);
 
   return (
-    <RecoilRoot>
-      <QueryClientProvider client={queryClient}>
-        <ErrorBoundary>
-          <ChakraProvider theme={theme}>
-            <CSSReset />
-            {
-              !mounted ? <AppLoading /> : (
-                <AppLayout >
-                  <Component {...pageProps} />
-                </AppLayout>
-              )
-            }
-          </ChakraProvider>
-          <Toaster position="top-right" toastOptions={{ duration: 10000 }} />
-        </ErrorBoundary>
-      </QueryClientProvider>
-    </RecoilRoot>
+    <>
+      <Head>
+        <link rel="shortcut icon" href="/favicon.ico" />
+      </Head>
+      <RecoilRoot>
+        <QueryClientProvider client={queryClient}>
+          {/* <ErrorBoundary> */}
+            <ChakraProvider theme={theme}>
+              <CSSReset />
+              {
+                !mounted ? <AppLoading /> : (
+                  <AppLayout >
+                    <Component {...pageProps} />
+                  </AppLayout>
+                )
+              }
+            </ChakraProvider>
+            <Toaster position="top-right" toastOptions={{ duration: 10000 }} />
+          {/* </ErrorBoundary> */}
+        </QueryClientProvider>
+      </RecoilRoot>
+    </>
   )
 }
