+++ pages/pools/index.tsx
@@ -1,12 +1,5 @@
 import Pools from 'components/Pages/Pools'
-import Page from 'components/Page';
 
-const PoolsPage = () => {
-  return (
-    <Page>
-      <Pools />
-    </Page>
-  )
-}
+const PoolsPage = () => (<Pools />)
 
 export default PoolsPage
