+++ pages/pools/manage_liquidity.tsx
@@ -1,0 +1,5 @@
+import ManageLiquidity from 'components/Pages/ManageLiquidity'
+
+const ManageLiquidityPage = () => ( <ManageLiquidity />)
+
+export default ManageLiquidityPage
