--- pages/pools/managelp.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import ManageLP from 'components/Pages/ManageLP'
-
-const ManageLPPage = () => {
-  return (
-    <ManageLP />
-  )
-}
-
-export default ManageLPPage
