+++ pages/pools/new_position.tsx
@@ -1,0 +1,5 @@
+import NewPosition from 'components/Pages/NewPosition'
+
+const NewPositionPage = () => (<NewPosition />)
+
+export default NewPositionPage
