--- pages/pools/providelp.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import ProvideLP from 'components/Pages/ProvideLP'
-
-const ProvideLPPage = () => {
-  return (
-    <ProvideLP />
-  )
-}
-
-export default ProvideLPPage
