+++ queries/querySwapInfo.ts
@@ -2,10 +2,6 @@ import { getSwapInfo } from '../services/swap'
 
 export async function querySwapInfo({ context: { client }, swap_address }) {
 
-  console.log({
-    client, swap_address
-  })
-
   const swap = await getSwapInfo(swap_address, client)
   const [asset1, asset2] = swap?.assets || []
   return {
